requirements.txt
# Core LangChain and LangGraph packages (Latest 2025 versions)
langchain>=0.3.0
langgraph>=0.5.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchainhub
langsmith>=0.2.0

# Enhanced search and web scraping
duckduckgo-search
beautifulsoup4
requests
lxml
html2text

# Google Custom Search Engine API
google-api-python-client>=2.0.0

# UI and visualization
streamlit>=1.28.0
gradio
matplotlib
plotly
mermaid-py

# Data processing and analysis
pandas
numpy
python-dotenv

# Performance and error handling
tiktoken
pydantic>=2.0.0
retry
tenacity

# Legal document processing
pypdf2
python-docx
pdfplumber

# Bulgarian text processing
regex
unidecode

# Logging and monitoring
structlog
rich

# Optional: Advanced features
# pygraphviz  # For advanced graph visualization (requires system dependencies)
# opencv-python  # For image processing if needed